Revert "Refactor/styles cleanup"
