+++ .todo
@@ -10,13 +10,12 @@ Filaments:
     ☐ add editing/deleting filaments @medium
 
 UI/UX:
-    ✔ auto-update quantity after printing @critical @done (9/10/2025, 6:07:36 PM)
+    ☐ auto-update quantity after printing @critical 
     ☐ improve accessibility (ARIA labels, keyboard navigation) @medium
     ☐ printing settings order @medium
-    ✔ dark mode needs work @low @done (9/10/2025, 5:48:57 PM)
+    ☐ dark mode needs work @low
     ☐ add confirmation modals for deletions @low
     ☐ improve loading states (spinners, skeletons) @low
-    ☐ cool quantity animation @low
 
 Firebase:
     ☐ add notes to Firebase @high
@@ -25,7 +24,7 @@ Firebase:
 
 History:
     ☐ add History of changes @high
-    ☐ add printing History @low
+    ☐ add printing History @high
 
 Code Quality:
     ☐ two confige files is dumb @high
