+++ src/pages/FilamentDetails.css
@@ -5,7 +5,7 @@
 }
 
 .back-link {
-  color: var(--accent);
+  color: #60a5fa;
   text-decoration: none;
 }
 
@@ -46,23 +46,23 @@
 }
 
 .quantity.superior {
-  color: var(--quantity-superior);
+  color: #96eaff; /* niebieski */
   font-weight: 600;
 }
 .quantity.high {
-  color: var(--quantity-high);
+  color: #4ade80; /* zielony */
   font-weight: 600;
 }
 .quantity.medium {
-  color: var(--quantity-medium);
+  color: #facc15; /* żółty */
   font-weight: 600;
 }
 .quantity.low {
-  color: var(--quantity-low);
+  color: #f87171; /* czerwony */
   font-weight: 600;
 }
 .quantity.extremlyLow {
-  color: var(--quantity-extremlyLow);
+  color: #f73898; /* czerwony */
   font-weight: 600;
 }
 
@@ -86,7 +86,7 @@
 }
 
 .settings-card {
-  background: var(--card-bg);
+  background: #1f2937;
   border-radius: 0.75rem;
   box-shadow: 0 2px 6px rgba(0, 0, 0, 0.4);
   padding: 1rem;
@@ -121,11 +121,11 @@
 }
 
 .setting-key {
-  color: var(--text);
+  color: #d1d5db;
 }
 
 .setting-value {
-  background: var(--setting-value-bg);
+  background: #374151;
   padding: 0.25rem 0.5rem;
   border-radius: 0.25rem;
   cursor: pointer;
@@ -133,20 +133,18 @@
 }
 
 .setting-value:hover {
-  background: var(--setting-value-bg-hover);
+  background: #4b5563;
 }
 
 .tooltip {
   position: absolute;
   top: -1.5rem;
   right: 0;
-  background: var(--bg);
-  box-shadow: 0px 2px 8px rgba(3, 7, 18, 0.03),
-    0px 8px 31px rgba(3, 7, 18, 0.05), 0px 17px 70px rgba(3, 7, 18, 0.08);
+  background: #111827;
   padding: 0.25rem 0.5rem;
   border-radius: 0.25rem;
   font-size: 0.75rem;
-  color: var(--text);
+  color: #f9fafb;
   white-space: nowrap;
 }
 
@@ -172,9 +170,9 @@
   height: 8rem;
   padding: 0.5rem;
   border-radius: 0.5rem;
-  background: var(--card-bg);
-  color: var(--text);
-  border: 1px solid var(--border);
+  background: #1f2937;
+  color: #e5e7eb;
+  border: 1px solid #374151;
 }
 
 .save-button {
